requirements.txt
pytrends==4.9.2
openai==1.59.5
python-dotenv==1.0.1
typer==0.15.1
rich==13.9.4
requests==2.32.3
sentence-transformers==3.3.1
hdbscan==0.8.40
faiss-cpu==1.9.0.post1
